consolidated propose beacon block tests and improved code coverage
