--- validator/client/beacon-api/propose_beacon_block_electra_test.go
@@ -1,49 +1,0 @@
-package beacon_api
-
-import (
-	"bytes"
-	"encoding/json"
-	"testing"
-
-	"github.com/OffchainLabs/prysm/v6/api/server/structs"
-	rpctesting "github.com/OffchainLabs/prysm/v6/beacon-chain/rpc/eth/shared/testing"
-	"github.com/OffchainLabs/prysm/v6/testing/assert"
-	"github.com/OffchainLabs/prysm/v6/testing/require"
-	"github.com/OffchainLabs/prysm/v6/validator/client/beacon-api/mock"
-	"go.uber.org/mock/gomock"
-)
-
-func TestProposeBeaconBlock_Electra(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	jsonRestHandler := mock.NewMockJsonRestHandler(ctrl)
-
-	var blockContents structs.SignedBeaconBlockContentsElectra
-	err := json.Unmarshal([]byte(rpctesting.ElectraBlockContents), &blockContents)
-	require.NoError(t, err)
-	genericSignedBlock, err := blockContents.ToGeneric()
-	require.NoError(t, err)
-
-	electraBytes, err := json.Marshal(blockContents)
-	require.NoError(t, err)
-	// Make sure that what we send in the POST body is the marshalled version of the protobuf block
-	headers := map[string]string{"Eth-Consensus-Version": "electra"}
-	jsonRestHandler.EXPECT().Post(
-		gomock.Any(),
-		"/eth/v2/beacon/blocks",
-		headers,
-		bytes.NewBuffer(electraBytes),
-		nil,
-	)
-
-	validatorClient := &beaconApiValidatorClient{jsonRestHandler: jsonRestHandler}
-	proposeResponse, err := validatorClient.proposeBeaconBlock(t.Context(), genericSignedBlock)
-	assert.NoError(t, err)
-	require.NotNil(t, proposeResponse)
-
-	expectedBlockRoot, err := genericSignedBlock.GetElectra().Block.HashTreeRoot()
-	require.NoError(t, err)
-
-	// Make sure that the block root is set
-	assert.DeepEqual(t, expectedBlockRoot[:], proposeResponse.BlockRoot)
-}
